chore: move reders of components sections to a specific folder
