+++ src/components/docs/component-sections-renders/properties-section.tsx
